[MOD-8039] Reorganize stats files (#5329)
* new file : info/INFO_MODULES

responsible to expose all aggregated shard stats to INFO command

move RS_moduleInfoFunc from module-init to

move all add to info functions to a dedicated file:

from global_stats.c
FieldsGlobalStats_AddToInfo
TotalGlobalStats_Queries_AddToInfo
DialectsGlobalStats_AddToInfo
IndexesGlobalStats_AddToInfo

from config.c:
RSConfig_AddToInfo

introduce FieldsGlobalStats_GetIndexError to expose fields errors
move queries stats from TotalGlobalStats to seperate struct QueriesGlobalStats

* rediSearchAPI::RediSearch_TotalInfo calls IndexesInfo_TotalInfo from indexes_info

IndexesInfo_TotalInfo (orignally  rediSearchAPI::RediSearch_TotalInfo) calls IndexSpec_TotalMemUsage instead of rediSearchAPI::RediSearch_TotalMemUsage

introduce indexes_info
rename TotalSpecsFieldInfo and  TotalSpecsInfo:
specs->indexes
move TotalIndexesFieldsInfo and TotalIndexesInfo from info_command to indexes_info

move global stats to info dir
move info_command to info dir

introduce util/units. A file for all macros handling shared units conversion
remove CLOCKS_PER_MILLISEC from profile.h, include util/units.h instead

add a cpp test for llapi indexes info

* release index
remove unused header from module-init

* remove TODO

* review fixes

rename INFO_MODULES->info_redis

rename MEMORY_HUMAN->MEMORY_MB

(cherry picked from commit 175d78713598207192fd4e26b41fd88f4a32ae83)
