+++ src/module.c
@@ -38,12 +38,13 @@
 #include "alias.h"
 #include "module.h"
 #include "rwlock.h"
-#include "info_command.h"
+#include "info/info_command.h"
 #include "rejson_api.h"
 #include "geometry/geometry_api.h"
 #include "reply.h"
 #include "resp3.h"
-#include "global_stats.h"
+#include "info/global_stats.h"
+#include "util/units.h"
 
 
 /* FT.MGET {index} {key} ...
