+++ src/redisearch_api.h
@@ -10,7 +10,7 @@
 #include "redismodule.h"
 #include <limits.h>
 #include "fork_gc.h"
-#include "info_command.h"
+#include "info/indexes_info.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -208,7 +208,7 @@ MODULE_API_FUNC(RSFieldID, RediSearch_CreateField)
   RediSearch_CreateField(idx, name, RSFLDTYPE_GEO, RSFLDOPT_NONE)
 #define RediSearch_CreateVectorField(idx, name) \
   RediSearch_CreateField(idx, name, RSFLDTYPE_VECTOR, RSFLDOPT_NONE)
-// TODO: GEOMETRY 
+// TODO: GEOMETRY
 // #define RediSearch_CreateGeometryField(idx, name) \
 //   RediSearch_CreateField(idx, name, RSFLDTYPE_GEOMETRY, RSFLDOPT_NONE)
 
@@ -360,7 +360,7 @@ MODULE_API_FUNC(size_t, RediSearch_MemUsage)(RSIndex* sp);
 
 MODULE_API_FUNC(size_t, RediSearch_TotalMemUsage)(void);
 
-MODULE_API_FUNC(TotalSpecsInfo, RediSearch_TotalInfo)(void);
+MODULE_API_FUNC(TotalIndexesInfo, RediSearch_TotalInfo)(void);
 
 MODULE_API_FUNC(InfoGCStats, RediSearch_GC_total)(void);
